% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\name{abt}
\alias{abt}
\title{Antibiotic time course experiment.}
\format{A phyloseq object.}
\description{
A phyloseq object describing an experiment on three people who took
two courses of cipro. Reference is Dethlefsen and Relman, PNAS
(2010), at https://www.ncbi.nlm.nih.gov/pubmed/20847294
}
