% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\name{abt_data}
\alias{abt_data}
\title{Antibiotic time course experiment, alternate format}
\format{A list with four components.}
\description{
A list with components for the normalized OTU abundances (otutab),
the covariance matrix for the tree relating the OTUs (Q), data
about the samples (sampledata), and taxonomic information about the
OTUs (variabledata).
}
