% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptivegpca-package.R
\name{adaptivegPCA}
\alias{adaptivegPCA}
\title{Adaptive gPCA}
\usage{
adaptivegPCA(X, Q, weights = rep(1, nrow(X)), k = 2)
}
\arguments{
\item{X}{A data matrix}

\item{Q}{The inner product matrix on the rows of X, can be given as
an eigendecomposition (formatted as the output from eigen()).}

\item{weights}{Data weights for the rows of X.}

\item{k}{The number of components to return.}
}
\description{
Performs adaptive generalized PCA
}

