% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptivegpca-package.R
\name{adaptivegPCA2}
\alias{adaptivegPCA2}
\title{Adaptive gPCA}
\usage{
adaptivegPCA2(X, Q, weights = rep(1, nrow(X)), k = 2)
}
\arguments{
\item{X}{An n x p matrix with rows corresponding to observations.}

\item{Q}{A p x p psd matrix giving the structure of the variables.}

\item{weights}{Data weights for the rows of X.}

\item{k}{The number of components to return.}
}
\description{
Performs adaptive gPCA where the noise can be non-spherical
(identity plus a scalar multiple of Q^(-1)).
}

