% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\name{AntibioticPhyloseq}
\alias{AntibioticPhyloseq}
\title{Antibiotic time course experiment.}
\format{A phyloseq object.}
\description{
A phyloseq object describing a time course experiment in which
three people two courses of cipro and had their gut microbiomes
sampled. See Dethlefsen and Relman, PNAS (2010), at
https://www.ncbi.nlm.nih.gov/pubmed/20847294 for more details.
}
