% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\name{AntibioticSmall}
\alias{AntibioticSmall}
\title{A subset of the antibiotic data}
\format{A list with three components.}
\description{
This is a smaller version of the \code{AntibioticPhyloseq} dataset,
for use in the examples so that the running time isn't so long. It
has the same samples and a randomly selected set of 200 of the
taxa. It is stored as a list with three components: the normalized
OTU abundances (\code{X}), the similarity matrix for the taxa
(\code{Q}), and the diagonal weight matrix (\code{D}, the identity
matrix).
}
