% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptivegpca-package.R
\name{check_axes}
\alias{check_axes}
\title{Check that the axes specified are valid}
\usage{
check_axes(axes, x)
}
\arguments{
\item{axes}{A set of user-specified axes.}

\item{x}{Object of class \code{adaptivegpca}.}
}
\description{
Check that the axes specified are valid
}
\keyword{internal}
