% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgets.R
\name{check_phyloseq}
\alias{check_phyloseq}
\title{Check compatibility of agpca and phyloseq objects}
\usage{
check_phyloseq(agpcafit, physeq)
}
\arguments{
\item{agpcafit}{An adaptivegpca object.}

\item{physeq}{A phyloseq object.}
}
\description{
Check that the dimensions of the agpca object match the phyloseq
object and that the phyloseq object has a taxonomy table and a
phylogenetic tree.
}
\keyword{internal}
