% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varianceComponents.R
\name{estimateComponents}
\alias{estimateComponents}
\title{Estimate parameters in hierarchical model}
\usage{
estimateComponents(X, Q, maxit = 8, tol = 10^(-10), plot = FALSE,
  Qeig = NULL)
}
\arguments{
\item{X}{A n x p matrix with n samples in p-dimensional space.}

\item{Q}{The prior variance on the means.}

\item{maxit}{The maximum number of bisections to try.}

\item{tol}{Stop if the absolute value of the derivative is less than this. œ}
}
\value{
A list with r and sigma.
}
\description{
Estimates the values of r and sigma in a model X sim N(0, sigma^2 (r Q + (1 - r)I)) using bisection.
}

