% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varianceComponents.R
\name{estimateComponents2}
\alias{estimateComponents2}
\title{Estimate variance components}
\usage{
estimateComponents2(X, Q)
}
\arguments{
\item{X}{An n x p matrix with rows corresponding to observations.}

\item{Q}{A p x p psd matrix giving the structure.}
}
\value{
A vector with r1 and r2
}
\description{
Estimate variance components in a two-parameter model where X ~
N(0, sigma^2 (r1 Q + (1 - r1) (r2 Q^(-1) + (1 - r2) I)))
}

