% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptivegpca-package.R
\name{findReflection}
\alias{findReflection}
\title{Find reflection}
\usage{
findReflection(df1, df2)
}
\arguments{
\item{df1}{The base data frame.}

\item{df2}{The data frame that will be reflected across either the
x-axis or the y-axis (or both or neither) so that the points in it
most closely match df1.}
}
\value{
A vector of length \code{ncol(df1)}: Multiplying the first
column of df2 by the first element and multiplying the second
column of df2 by the second element and so on gives the optimal
reflection.
}
\description{
Find a reflection of one data frame so that it most closely matches
another.
}
\keyword{internal}
