% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptivegpca-package.R
\name{gPCA}
\alias{gPCA}
\title{gPCA}
\usage{
gPCA(X, Q, D = rep(1, nrow(X)), k)
}
\arguments{
\item{X}{A data matrix}

\item{Q}{An inner product matrix for the rows, either as a matrix
or an eigendecomposition.}

\item{D}{An inner product matrix for the columns.}

\item{k}{The number of components to return.}
}
\value{
A list with principal components, scores, and variable positions.
}
\description{
Performs standard gPCA with k components on a data matrix X with
row inner product Q and column inner product D.
}
