% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptivegpca-package.R
\name{gpcaEvecs}
\alias{gpcaEvecs}
\title{Performs gPCA with pre-computed eigenvectors and eigenvalues}
\usage{
gpcaEvecs(X, evecs, evals, D = rep(1, nrow(X)), k)
}
\arguments{
\item{X}{Data matrix.}

\item{evecs}{Eigenvectors of Q}

\item{evals}{Eigenvalues of Q}

\item{D}{Inner product matrix for the columns.}

\item{k}{The number of components to return.}
}
\description{
Performs gPCA with pre-computed eigenvectors and eigenvalues
}

