% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptivegpca-package.R
\name{gpcaFullFamily}
\alias{gpcaFullFamily}
\title{Make a sequence of ordinations}
\usage{
gpcaFullFamily(X, Q, weights = rep(1, nrow(X)), k = 2, rvec = (0:100)/100,
  findReflections = TRUE, returnLong = FALSE, sampledata = NULL,
  variabledata = NULL)
}
\arguments{
\item{X}{A data matrix.}

\item{Q}{An inner product on the rows of X.}

\item{weights}{A vector of weights for the rows of X.}

\item{k}{The number of components to compute for each ordination.}

\item{rvec}{The values of r for which to make the ordinations.}

\item{findReflections}{Whether or not flip the axes so as to make
neighboring ordinations as close as possible. If k is very large
this should be false since all possible axis combinations are
searched over.}

\item{returnLong}{Return a long data frame with the
samples/variables instead of a list of data frames.}

\item{sampledata}{Extra sample data to be included along with the
sample scores.}

\item{variabledata}{Extra variable data to be included along with
the variable loadings.}
}
\value{
A list with three components: one holding the location
points, one holding the species points, and one holding the
variance fractions. Each list is itself a list of data frames
(location/species points) or of vectors (for the variances).
}
\description{
Creates a sequence of gPCA data representations. One end of the
sequence doesn't take into account the variable structure at all
(PCA), and the other weights the structure of the variables very
highly.
}
