% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varianceComponents.R
\name{gradLik}
\alias{gradLik}
\title{Derivative of the likelihood}
\usage{
gradLik(Xtilde, r, D)
}
\arguments{
\item{Xtilde}{The transformed data}

\item{r}{r}

\item{D}{The eigenvalues of Q}
}
\description{
Derivative of the likelihood in the hierarchical model as a
function of \eqn{r}.
}
\keyword{internal}
