% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varianceComponents.R
\name{likelihood}
\alias{likelihood}
\title{The likelihood at a given value of \eqn{r} and \eqn{\sigma}}
\usage{
likelihood(Xtilde, sigma, r, D)
}
\arguments{
\item{Xtilde}{The transformed data}

\item{sigma}{Overall scaling factor.}

\item{r}{r}

\item{D}{The eigenvalues of Q}
}
\description{
The likelihood at a given value of \eqn{r} and \eqn{\sigma}
}
\keyword{internal}
