% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varianceComponents.R
\name{likelihood_two_params}
\alias{likelihood_two_params}
\title{Likelihood of data in two-parameter model}
\usage{
likelihood_two_params(r1, r2, Xtilde, D)
}
\arguments{
\item{r1}{Coefficient of Q}

\item{r2}{Coefficient of Q^(-1) in the noise part.}

\item{Xtilde}{The data projected onto the eigenvectors of Q.}

\item{D}{The eigenvalues of Q}
}
\value{
The marginal likelihood of the data given r1 and r2.
}
\description{
Likelihood of data in two-parameter model
}
\keyword{internal}
