% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varianceComponents.R
\name{likelihoodR}
\alias{likelihoodR}
\title{The likelihood at a given value of \eqn{r} and the maximizing \eqn{\sigma} for
that value of \eqn{r}.}
\usage{
likelihoodR(Xtilde, r, D)
}
\arguments{
\item{Xtilde}{The transformed data}

\item{r}{r}

\item{D}{The eigenvalues of Q}
}
\description{
The likelihood at a given value of \eqn{r} and the maximizing \eqn{\sigma} for
that value of \eqn{r}.
}
\keyword{internal}
