% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptivegpca-package.R
\name{normalizeMatrix}
\alias{normalizeMatrix}
\title{Normalizes a matrix.}
\usage{
normalizeMatrix(X)
}
\arguments{
\item{X}{The matrix to be normalized.}
}
\value{
A list with the normalized matrix and the corresponding metric.
}
\description{
Normalizes a count matrix and returns the corresponding metric.
}

