% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptivegpca-package.R
\name{print.adaptivegpca}
\alias{print.adaptivegpca}
\title{Print an adaptivegpca object}
\usage{
\method{print}{adaptivegpca}(x, ...)
}
\arguments{
\item{x}{\code{adaptivegpca} object.}

\item{...}{Not used.}
}
\description{
Print an adaptivegpca object
}
