% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptivegpca-package.R
\name{processPhyloseq}
\alias{processPhyloseq}
\title{Make the input matrices for PCA}
\usage{
processPhyloseq(physeq, ca = FALSE)
}
\arguments{
\item{physeq}{A phyloseq object.}

\item{ca}{If TRUE, do the normalization as for correspondence
analysis (transform counts to relative abundances, compute sample
weights, center the relative abundances according to the sample
weights). Otherwise, simply center the data.}
}
\value{
A list of the matrix to perform gPCA on (X), the norm for
the rows (Q), and the sample weights (weights).
}
\description{
Takes a phyloseq object and creates the matrices necessary to do a
generalized PCA.
}
