% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptivegpca-package.R
\name{processPhyseq}
\alias{processPhyseq}
\title{Make the input matrices for PCA}
\usage{
processPhyseq(physeq)
}
\arguments{
\item{physeq}{A phyloseq object.}
}
\value{
A list of the matrix to perform gPCA on (X), the norm for
the rows (sigma), and the norm for the columns (D).
}
\description{
Takes a phyloseq object and (optionally) some constraints and
creates the matrices necessary to do a generalized PCA.
}

