% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varianceComponents.R
\name{sigma2OfR}
\alias{sigma2OfR}
\title{Value of sigma^2 that maximizes the likelihood for a given value of
r}
\usage{
sigma2OfR(Xtilde, r, D)
}
\arguments{
\item{Xtilde}{The transformed data}

\item{r}{r}

\item{D}{The eigenvalues of Q}
}
\description{
Value of sigma^2 that maximizes the likelihood for a given value of
r
}

