% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgets.R
\name{visualizeFullFamily}
\alias{visualizeFullFamily}
\title{Shiny gadget for adaptive gPCA}
\usage{
visualizeFullFamily(fullFamily, sampleData = NULL, sample_mapping = aes(x =
  X1, y = X2), varData = NULL, var_mapping = aes(x = X1, y = X2))
}
\arguments{
\item{fullFamily}{The output from gpcaFullFamily}

\item{sampleData}{Optional data used for plotting the samples}

\item{sample_mapping}{An aesthetic mapping to be passed to ggplot
for plotting the samples}

\item{varData}{Optional data used for plotting the variables}

\item{var_mapping}{An aesthetic mapping to be passed to ggplot for
plotting the variables}
}
\description{
Shiny gadget that shows the ordinations from an entire family of
gPCAs and returns a gPCA object with the one selected
}

