% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varianceComponents.R
\name{estimateComponents}
\alias{estimateComponents}
\title{Estimate parameters in hierarchical model}
\usage{
estimateComponents(X, Q, maxit = 8, tol = 10^(-10), Qeig = NULL)
}
\arguments{
\item{X}{A n x p matrix with n samples in p-dimensional space.}

\item{Q}{The prior variance on the means.}

\item{maxit}{The maximum number of bisections to try.}

\item{tol}{Stop if the absolute value of the derivative is less than this.}

\item{Qeig}{If the eigendecomposition of Q is already computed, it
can be included here.}
}
\value{
A list with r, sigma, and the values of the likelihood on a
grid of values between 0 and 1.
}
\description{
Estimates the values of r and sigma in a model X sim N(0, sigma^2
(r Q + (1 - r)I)).
}

