% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varianceComponents.R
\name{gradSigma2OfR}
\alias{gradSigma2OfR}
\title{Derivative of \eqn{\sigma^2}}
\usage{
gradSigma2OfR(Xtilde, r, D)
}
\arguments{
\item{Xtilde}{The transformed data}

\item{r}{r}

\item{D}{The eigenvalues of Q}
}
\description{
Derivative of \eqn{\sigma^2(r)} in the hierarchical model
}
\keyword{internal}
