% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varianceComponents.R
\name{varianceOnEvecs}
\alias{varianceOnEvecs}
\title{Variance along eigenvectors}
\usage{
varianceOnEvecs(X, Q)
}
\arguments{
\item{X}{The data, each row a sample.}

\item{Q}{The inner product matrix, either as a matrix or as its
eigendecomposition (the output from eigen).}
}
\value{
A vector containing the variances along each of the
eigenvectors.
}
\description{
Find the variance of the data along each of the eigenvectors
}

